+++ src/program/movie/MovieFileInputs.cpp
@@ -174,7 +174,7 @@ const AllInputs& MovieFileInputs::getInputs(uint64_t pos)
     // std::unique_lock<std::mutex> lock(input_list_mutex);
 
     if (pos >= input_list.size()) {
-        pos = input_list.size();
+        pos = input_list.size() - 1;
     }
 
     /* Special case for zero framerate */